feat: 세션 방식 수정
feat: 슈퍼 관리자 초기 데이터 생성 추가

feat: 관리자 상태 변경 API 추가

feat: 관리자 리스트 조회 API 추가

refactor: 클래스명 변경 및 정리

+++ src/main/java/com/ecommerce/backoffice/domain/admin/service/AdminService.java
@@ -3,6 +3,8 @@ package com.ecommerce.backoffice.domain.admin.service;
 
 import com.ecommerce.backoffice.domain.admin.dto.request.*;
 import com.ecommerce.backoffice.domain.admin.dto.response.*;
+import com.ecommerce.backoffice.domain.admin.dto.session.SessionAdmin;
+import com.ecommerce.backoffice.domain.admin.dto.session.SessionKey;
 import com.ecommerce.backoffice.domain.admin.dto.session.TimeProvider;
 import com.ecommerce.backoffice.domain.admin.entity.Admin;
 import com.ecommerce.backoffice.domain.admin.enums.AdminRole;
@@ -13,6 +15,10 @@ import com.ecommerce.backoffice.global.error.CommonException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -52,28 +58,50 @@ public class AdminService {
                 () -> new CommonException(CommonError.USER_NOT_FOUND)
         );
 
-        if(!passwordEncoder.matches(request.password(), admin.getPassword())){
+        if (!passwordEncoder.matches(request.password(), admin.getPassword())) {
             throw new CommonException(CommonError.INVALID_PASSWORD);
         }
 
         if (admin.getStatus() != AdminStatus.APPROVED) {
             switch (admin.getStatus()) {
-                case PENDING   -> throw new CommonException(CommonError.PENDING_ACCOUNT);
+                case PENDING -> throw new CommonException(CommonError.PENDING_ACCOUNT);
                 case SUSPENDED -> throw new CommonException(CommonError.SUSPENDED_ACCOUNT);
                 case INACTIVATE -> throw new CommonException(CommonError.INACTIVE_ACCOUNT);
-                default        -> throw new CommonException(CommonError.LOGIN_FAILED);
+                default -> throw new CommonException(CommonError.LOGIN_FAILED);
             }
         }
 
         HttpSession session = sessionRequest.getSession(true);
-        session.setAttribute("ADMIN_ID", admin.getId());
+        session.setAttribute("LOGIN_ADMIN", SessionAdmin.from(admin));
 
         return AdminLoginResponse.from(admin);
     }
 
+    // 쿼리 파라 미터 조회
+    @Transactional(readOnly = true)
+    public GetAdminListResponse getAllAdmins(HttpSession session, String keyword, Integer page, Integer size, String sortBy, String sortDir, String role, String status) {
+        SessionKey.checkSuperAdmin(session);
+        int safePage = (page == null || page < 1) ? 1 : page;
+        int safeSize = (size == null || size < 1) ? 10 : size;
+
+        Sort.Direction dir = "desc".equalsIgnoreCase(sortDir) ? Sort.Direction.DESC : Sort.Direction.ASC;
+        String safeSortBy = (sortBy == null || sortBy.isBlank()) ? "createdAt" : sortBy;
+
+        AdminRole roleEnum = (role == null || role.isBlank()) ? null : AdminRole.valueOf(role);
+        AdminStatus statusEnum = (status == null || status.isBlank()) ? null : AdminStatus.valueOf(status);
+
+        Pageable pageable = PageRequest.of(safePage - 1, safeSize, Sort.by(dir, safeSortBy));
+
+        Page<Admin> pageResult = adminRepository.findAllAdmins(keyword, roleEnum, statusEnum, pageable);
+
+        return GetAdminListResponse.from(pageResult);
+    }
+
+
     // 관리자 상세 조회
     @Transactional(readOnly = true)
-    public GetAdminDetailResponse getOne(Long adminId) {
+    public GetAdminDetailResponse getOne(HttpSession session, Long adminId) {
+        SessionKey.checkSuperAdmin(session);
         Admin admin = findById(adminId);
         return GetAdminDetailResponse.from(admin);
     }
@@ -87,9 +115,11 @@ public class AdminService {
 
     // 관리자 정보 수정
     @Transactional
-    public UpdateAdminResponse updateAdminInfo(Long adminId, UpdateAdminRequest requestBody) {
-        Admin admin = adminRepository.findById(adminId)
-                .orElseThrow(() -> new CommonException(CommonError.ADMIN_NOT_FOUND));
+    public UpdateAdminResponse updateAdminInfo(HttpSession session, Long adminId, UpdateAdminRequest requestBody) {
+        SessionKey.checkSuperAdmin(session);
+        Admin admin = adminRepository.findById(adminId).orElseThrow(
+                () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
+        );
 
         boolean hasAny = (requestBody.name() != null) || (requestBody.email() != null) || (requestBody.phone() != null);
 
@@ -115,17 +145,32 @@ public class AdminService {
 
     // 관리자 역할 변경
     @Transactional
-    public UpdateAdminRoleResponse updateAdminRole(Long adminId, UpdateAdminRoleRequest requestBody) {
+    public UpdateAdminRoleResponse updateAdminRole(HttpSession session, Long adminId, UpdateAdminRoleRequest requestBody) {
+        SessionKey.checkSuperAdmin(session);
         Admin admin = adminRepository.findById(adminId).orElseThrow(
                 () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
         );
-        admin.updateAdminRole(requestBody.role().getRole());
+        admin.updateAdminRole(requestBody.role());
         return new UpdateAdminRoleResponse(admin.getId(), admin.getRole());
+    }
+
+    // 관리자 상태 변경
+    @Transactional
+    public PatchAdminStatusChangeResponse changeStatus(HttpSession session, Long adminId, PatchAdminStatusChangeRequest requestBody) {
+        SessionKey.checkSuperAdmin(session);
+        Admin admin = adminRepository.findById(adminId)
+                .orElseThrow(
+                        () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
+                );
+
+        admin.adminStatus(requestBody.status());
+        return PatchAdminStatusChangeResponse.from(admin);
     }
 
     // 관리자 삭제
     @Transactional
-    public void deleteAdmin(Long adminId) {
+    public void deleteAdmin(HttpSession session, Long adminId) {
+        SessionKey.checkSuperAdmin(session);
         Admin admin = adminRepository.findById(adminId)
                 .orElseThrow(() -> new CommonException(CommonError.ADMIN_NOT_FOUND));
         adminRepository.delete(admin);
@@ -133,7 +178,8 @@ public class AdminService {
 
     // 관리자 승인
     @Transactional
-    public DecisionAdminResponse approveAdmin(Long adminId) {
+    public PatchDecisionAdminResponse approveAdmin(HttpSession session, Long adminId) {
+        SessionKey.checkSuperAdmin(session);
         Admin admin = adminRepository.findById(adminId)
                 .orElseThrow(
                         () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
@@ -142,12 +188,13 @@ public class AdminService {
             throw new CommonException(CommonError.ADMIN_NOT_PENDING);
         }
         admin.approve(timeProvider.now());
-        return DecisionAdminResponse.from(admin);
+        return PatchDecisionAdminResponse.from(admin);
     }
 
     // 관리자 거부
     @Transactional
-    public DecisionAdminResponse rejectAdmin(Long adminId, RejectAdminRequest requestBody) {
+    public PatchDecisionAdminResponse rejectAdmin(HttpSession session, Long adminId, PatchRejectAdminRequest requestBody) {
+        SessionKey.checkSuperAdmin(session);
         Admin admin = adminRepository.findById(adminId)
                 .orElseThrow(
                         () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
@@ -158,12 +205,12 @@ public class AdminService {
         }
 
         admin.reject(timeProvider.now(), requestBody.reason());
-        return DecisionAdminResponse.from(admin);
+        return PatchDecisionAdminResponse.from(admin);
     }
 
     // 내 프로필 조회
     @Transactional(readOnly = true)
-    public GetProfileResponse getProfile(Long id) {
+    public GetProfileResponse getProfile(HttpSession session, Long id) {
         Admin admin = adminRepository.findById(id).orElseThrow(
                 () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
         );
@@ -172,11 +219,14 @@ public class AdminService {
 
     // 내 프로필 수정
     @Transactional
-    public UpdateProfileResponse updateProfile(Long id, UpdateProfileRequest requestBody) {
+    public UpdateProfileResponse updateProfile(HttpSession session, Long id, UpdateProfileRequest requestBody) {
         Admin admin = adminRepository.findById(id)
-                .orElseThrow(() -> new CommonException(CommonError.ADMIN_NOT_FOUND));
+                .orElseThrow(
+                        () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
+                );
 
         boolean hasAny = (requestBody.name() != null) || (requestBody.email() != null) || (requestBody.phone() != null);
+
         if (!hasAny) {
             throw new CommonException(CommonError.INVALID_UPDATE_REQUEST);
         }
@@ -198,7 +248,7 @@ public class AdminService {
 
     // 내 비밀 번호 변경
     @Transactional
-    public UpdatePasswordResponse changePassword(Long id, UpdatePasswordRequest requestBody) {
+    public UpdatePasswordResponse changePassword(HttpSession session, Long id, UpdatePasswordRequest requestBody) {
 
         Admin admin = adminRepository.findById(id).orElseThrow(
                 () -> new CommonException(CommonError.ADMIN_NOT_FOUND)
